+++ CHANGELOG.md
@@ -21,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support animation playback with matched images in multi-panel view ([#1860](https://github.com/CARTAvis/carta-frontend/issues/1860)).
 * Modified spectral matching type selector in Image List Settings won't affect global preferences ([#2098](https://github.com/CARTAvis/carta-frontend/issues/2098)).
 * Enabled enter key to trigger filtering in the filter fields, and modified the buttons in the spectral line query widget and catalog widget ([#2046](https://github.com/CARTAvis/carta-frontend/issues/2046)).
-* Switched the Splatalogue URL from test to production ([#2238](https://github.com/CARTAvis/carta-frontend/issues/2238))
+* Switched the Splatalogue URL from test to production ([#2238](https://github.com/CARTAvis/carta-frontend/issues/2238)).
+* Changed image fitting center position error unit from arcsecond to second when the format is H:M:S; added integrated flux information in the fitting result.
 ### Fixed
 * Fixed bug where line region computation width cannot be changed in spatial profile setting widget ([#2000](https://github.com/CARTAvis/carta-frontend/issues/2000)).
 * Fixed when multiple images are open, PV generator can only produce PV preview with live update for one of the images ([#2171](https://github.com/CARTAvis/carta-frontend/issues/2171)).